fix: move functions to thanks file
